# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ovrly API",
    "version": "0.2.0-draft",
    "summary": "Contract between the Android client and the backend.",
    "description": "Ovrly contracts 0.2.0-draft (pre-1.0). Shared-media investigation reads derive state from the intake job (queued/leased -> queued, running -> running, failed -> failed with PROCESSING_FAILED, cancelled/deleted -> cancelled). Capture progress uses GET /v1/captures/{capture_id}; media analysis and claim extraction are not implemented yet. Every schema is a reference into schemas/*.schema.json so the OpenAPI document, the JSON Schema fixtures and the Android models cannot drift from each other. Every error response uses the shared error shape; stage-specific error codes are listed per operation and must be SCREAMING_SNAKE_CASE. Paths listed here are those the backend serves today (BE-05, #19, including account linking from BC-D07, the owner-scoped job actions from BE-04 part 3, #75, and incremental capture intake from BE-06, #24) plus the report versions, explicit saves and POST /v1/voice/actions of BE-10 (#33; the voice contract is BE-13, #67). Reanalysis and the export payload are BE-10 too.",
    "license": {
      "name": "PolyForm Noncommercial 1.0.0",
      "identifier": "PolyForm-Noncommercial-1.0.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development helper (scripts/backend.sh). No hosted server is published by this document."
    }
...
  "security": [{ "bearer": [] }],
...
  "paths": {
...
    "/v1/jobs/{job_id}/cancel": {
      "parameters": [{ "$ref": "#/components/parameters/JobId" }],
      "post": {
        "tags": ["jobs"],
        "operationId": "cancelJob",
        "summary": "Cancel one owned job: effective now or requested from its worker.",
        "description": "No request body; a body is rejected with VALIDATION_FAILED. A queued or already cancelled job answers 200 with cancellation effective; a leased or running job answers 202 with cancellation requested and the worker makes it effective. The first outcome is stored and replayed unchanged on repeats, even after the worker acknowledges or the API restarts: this is a receipt, not a progress snapshot, and it promises nothing about provider interruption or billing. Missing, other-owner and legacy ownerless jobs all answer the same 404.",
        "responses": {
          "200": {
            "description": "Cancellation effective.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobCancelResponse" }
              }
            }
          },
          "202": {
            "description": "Cancellation requested; the worker observes it at its next heartbeat or lease expiry.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobCancelResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientIdentityRejected" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "JOB_NOT_CANCELLABLE: the job is already published or failed.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/DatabaseUnavailable" }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "parameters": [{ "$ref": "#/components/parameters/JobId" }],
      "delete": {
        "tags": ["jobs"],
        "operationId": "deleteJob",
        "summary": "Tombstone one owned job and clear its database payload and result.",
        "description": "No request body; a body is rejected with VALIDATION_FAILED. Deletion revokes access, fences late worker publication and removes the job payload and published result in one transaction; the tombstone keeps ownership, fencing and provider-request metadata so a late callback is routed to nothing. cleanup_scope is job_payload_and_result only: uploads, provider copies, external artifacts and backups are not covered (#77). Repeating DELETE returns the same receipt without another mutation; a cancellation receipt is no longer readable after deletion (404).",
        "responses": {
          "200": {
            "description": "Job tombstoned; cleanup of the database payload and result complete.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobDeleteResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientIdentityRejected" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/DatabaseUnavailable" }
        }
      }
    },
...
    "/v1/voice/actions": {
      "post": {
        "tags": ["voice"],
        "operationId": "executeVoiceAction",
        "summary": "Execute one allowlisted voice action against one owned target.",
        "description": "Contract fixed by BE-13 (#67); the endpoint is BE-10 (#33). A well-formed request_id and an action string of 1 to 64 characters with a name outside the allowlist answers 200 denied with VOICE_ACTION_UNSUPPORTED. For an allowlisted action, a body that fails schema validation (missing or mismatched target, extra fields such as confirmed) is 422 VALIDATION_FAILED in the shared error shape. Other in-contract denials answer 200 with result denied: VOICE_TARGET_NOT_FOUND for a missing target and, without revealing existence, another owner's target (the server does not emit VOICE_TARGET_NOT_OWNED); VOICE_ACTION_INVALID_STATE when the job state does not allow the action. queue_retry and queue_continue never change a job: they are accepted for a queued or running job and denied for a finished, failed or cancelled one, because the queue has no transition out of a terminal state. Every request writes one audit row with no transcript. Repeating a request_id replays the stored response without running the action again; reusing it for a different action or target is 409 IDEMPOTENCY_KEY_REUSED.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VoiceActionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted or denied; see result.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VoiceActionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientIdentityRejected" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "409": {
            "description": "IDEMPOTENCY_KEY_REUSED: the request_id was already used by this caller for a different action or target.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/DatabaseUnavailable" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque token minted by POST /v1/principals/guest (ovk_ prefix). Identity comes only from this credential; a user_id, owner_id or principal_id in a body, query or header is rejected with CLIENT_IDENTITY_REJECTED."
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Echoed from a well-formed client X-Request-Id (opaque_id syntax) or generated. Equals error.request_id on every error response.",
        "schema": { "$ref": "schemas/common.schema.json#/$defs/opaque_id" }
      }
    },
    "parameters": {
...
      "JobId": {
        "name": "job_id",
        "in": "path",
        "required": true,
        "description": "Canonical lowercase UUID of the job, as the queue stores it. The path parameters of both job actions are described by schemas/job-action-request.schema.json; neither operation takes a body.",
        "schema": { "$ref": "schemas/job-action-request.schema.json#/properties/job_id" }
      },
...
    },
    "responses": {
...
      "Unauthenticated": {
        "description": "AUTHENTICATION_REQUIRED (no credential) or INVALID_CREDENTIAL (malformed, unknown or revoked), with WWW-Authenticate.",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "WWW-Authenticate": { "schema": { "type": "string" } }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "NOT_FOUND: no such object visible to the caller. Another owner's object answers the same way.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ValidationFailed": {
        "description": "VALIDATION_FAILED (body, query or path invalid) or CLIENT_IDENTITY_REJECTED.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ClientIdentityRejected": {
        "description": "CLIENT_IDENTITY_REJECTED: a user_id, owner_id or principal_id arrived in the query string or an X-User-Id style header; identity comes only from the bearer credential.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "InternalError": {
        "description": "INTERNAL_ERROR: unexpected failure; no internals are revealed.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "DatabaseUnavailable": {
        "description": "DATABASE_UNAVAILABLE: retryable; action retry.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
...
      "Error": { "$ref": "schemas/error.schema.json" },
...
      "VoiceActionRequest": { "$ref": "schemas/voice-action-request.schema.json" },
      "VoiceActionResponse": { "$ref": "schemas/voice-action-response.schema.json" },
      "JobCancelResponse": { "$ref": "schemas/job-cancel-response.schema.json" },
      "JobDeleteResponse": { "$ref": "schemas/job-delete-response.schema.json" }
    }
  }
}